requirements.txt
django
django-rest-framework
djangorestframework-gis
gunicorn
psycopg2